+++ inst/doc/example.R
@@ -1,5 +1,0 @@
-## ------------------------------------------------------------------------
-data("ChickWeight")
-ChickWeight <- dplyr::mutate(ChickWeight, Time = Time+1)
-ChickWeight <- NULL
-